Replace transaction send check with a more robust one
The original transaction send check depended on the node relaying the transaction
back to Armory. However this behavior is not guaranteed and, due to changes in Core,
almost never happens.

The only guaranteed thing that will happen is that if the transaction is rejected for
any reason, a reject message will be sent. If no reject message is received and the connection
remains open and the node responsive, then the transaction was successfully sent.

The transaction send check mechanism has been replaced with one that waits for a reject, and
if it does not receive one, it will test the connection with a ping-pong.