[WIP] Overhaul macOS build process
- Completely change the macOS build process. No external libraries outside of the cppForSwig subdirectory are now built. Users will have to rely on brew and pip in order to install certain required third-party libraries.
- Fix up the notes as required.

The idea is to make macOS compilation more like Linux compilation and remove the need to maintain/modify third-party libraries. This will also make it far easier to transition Armory from Python 2 to Python 3, Qt 4 to Qt 5, etc., along with making deterministic builds far easier to obtain down the road.

This is a work-in-progress. Objective-C code isn't compiled (and related code is commented out of the Python code), the notes need to be finalized and checked against a fresh macOS VM, Armory needs to be run for awhile to see if it's stable, etc.