Fix GUI/client build issues
When disabling the GUI but enabling tests, the code still builds the GUI-specific library. Rearrange files to fully decouple the GUI-specific library from the unit tests.

Note that this might warrant a full refactor one day. As things stand, a lot of code could be in the GUI-specific or CLI-specific libraries, but must go into a common library for the sake of unit testing. This isn't optimal. But, it works for now.